feedback based changes for approval/refusal alert
