+++ submit-web/src/components/App/Submission/Modals/ApproveSubmissionModal.tsx
@@ -1,9 +1,6 @@
 import { Box, Typography } from "@mui/material";
-import WarningBox from "@/components/Shared/Layouts/WarningBox";
-import WarningAmberOutlinedIcon from "@mui/icons-material/WarningAmberOutlined";
-import { BCDesignTokens } from "epic.theme";
 import ConfirmationModal from "@/components/Shared/Modals/ConfirmationModal";
-import AlertModal from "@/components/Shared/Modals/AlertModal";
+import OpenUpdateRequestsAlertModal from "./OpenUpdateRequestsAlertModal";
 
 type ApproveSubmissionModalProps = {
   onConfirm: () => void;
@@ -21,41 +18,9 @@ const ApproveSubmissionModal = ({
   // Show alert modal when there are open update requests
   if (hasOpenUpdateRequests) {
     return (
-      <AlertModal
-        title="Cannot Approve Package"
+      <OpenUpdateRequestsAlertModal
         onClose={onCancel}
-        closeText="Close"
-        description={
-          <Box
-            sx={{
-              display: "flex",
-              flexDirection: "column",
-              gap: 2,
-              width: "520px",
-            }}
-          >
-            <WarningBox sx={{ p: 1.5 }}>
-              <Box sx={{ display: "flex", gap: 1.5 }}>
-                <WarningAmberOutlinedIcon
-                  sx={{ color: BCDesignTokens.supportBorderColorWarning }}
-                />
-                <Box>
-                  <Typography variant="body1" sx={{ fontWeight: 700 }}>
-                    Package will be Locked
-                  </Typography>
-                  <Typography variant="body2">
-                    Package cannot be approved while there are open update requests.
-                    The following sections have open update requests:{" "}
-                    <b>{openRequestSectionNames.join(", ")}</b>
-                  </Typography>
-                  <Typography variant="body2" sx={{ mt: 1 }}>
-                    Please accept or withdraw all update requests before approving this package.
-                  </Typography>
-                </Box>
-              </Box>
-            </WarningBox>
-          </Box>
-        }
+        openRequestSectionNames={openRequestSectionNames}
       />
     );
   }
